fixed rubocop issues. moved three functions from checklist_test model to a new data_criteria_selector module
